backend/requirements.txt
qrcode==7.3.1
SQLAlchemy==1.4.28
Flask==2.0.1
Flask-Cors==3.0.10
Werkzeug==2.0.2
pandas==1.3.4
Pillow==8.4.0
python-dotenv==0.19.2
matplotlib==3.5.1
SQLAlchemy-Utils==0.38.2
pytesseract==0.3.9
tesseract==0.1.3